# docs/frontdesk-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Front-Desk Operations API
  version: 2.0.0-draft
  description: >
    Domain API for a front-desk voice agent and the front-desk web application,
    reusable across domains through domain packs (healthcare first). Designed
    so that a conversational LLM needs at most three tools: one to find
    availability, one to manage a booking, one to answer general questions
    from approved answers. All business rules live here, never in the MCP
    layer or the prompt.
  contact:
    name: Front-desk platform workstream
servers:
  - url: https://{environment}.example-product.com/api/v1
    variables:
      environment:
        default: beta
...
security:
  - oauth2: []

# -----------------------------------------------------------------------------
# MCP TOOL MAPPING (read by the MCP builder; ignored by the REST server)
# -----------------------------------------------------------------------------
# Three tools. All are thin wrappers over `Agent` operations. The wrapper adds
# nothing but: header injection from call context, an outcome envelope that is
# already the REST response, and stripping of the injected headers from the
# tool's input schema.
...
paths:
...
  /agent/availability-search:
    post:
      tags: [Agent]
      operationId: agentAvailabilitySearch
      summary: Find resources, sessions and bookable slots from the caller's own words
      description: >
        The single discovery call. Accepts any combination of resource name,
        category, need text and time preference, all as spoken.
        RULE: run the safety check first (tenant red-flag lexicon); if it fires,
        return routing.action=TRANSFER_EMERGENCY and nothing else.
        RULE: resolve identity over the whole directory before applying any
        date or availability filter, so ambiguity survives.
        RULE: never return an empty list for a failure; set outcome=COULD_NOT_CHECK.
        RULE: resolve relative time words ("tomorrow", "next Tuesday", "evening")
        server-side in the facility timezone; the LLM must not do calendar math.
      x-mcp:
        expose: true
        tool: find_availability
        inject_headers: [X-Call-Id, X-Caller-Number]
      parameters:
        - $ref: '#/components/parameters/CallId'
        - $ref: '#/components/parameters/CallerNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvailabilitySearchRequest' }
            examples:
              namedResourceTomorrowEvening:
                summary: Named resource, relative date, day part, spoken in Kannada
                value:
                  utterance: "ನಾಳೆ ಸಂಜೆ ಡಾಕ್ಟರ್ ಗರಿಮಾ ಇರ್ತಾರಾ"
                  language: kn
                  resourceName: "Dr Garima"
                  when: { expression: "tomorrow evening" }
              needFirst:
                summary: Need only, ASR-damaged
                value:
                  utterance: "I need a zoologist appointment for kidney stones scanning"
                  language: en
                  needText: "kidney stones scanning"
                  when: { expression: "today" }
              categoryGender:
                summary: Category plus preference
                value:
                  utterance: "lady doctor for thyroid"
                  language: en
                  category: "thyroid doctor"
                  preferences: { gender: FEMALE }
                  when: { expression: "today" }
      responses:
        '200':
          description: Understood meaning, routing decision, sessions with bookable slots.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilitySearchResponse' }
              example:
                outcome: FOUND
                asOf: "2026-09-25T17:02:11+05:30"
                routing: { action: OFFER_SLOTS }
                understood:
                  resources:
                    - { resourceId: res_garima, name: "Dr. Garima", confidence: 0.94, matchedOn: NAME_PHONETIC }
                  categories: []
                  dates: { from: "2026-09-26", to: "2026-09-26" }
                  dayPart: EVENING
                clarification: null
                results:
                  - resource:
                      resourceId: res_garima
                      name: "Dr. Garima"
                      localizedNames: { kn: "ಡಾ. ಗರಿಮಾ", hi: "डॉ. गरिमा" }
                      categories: [ { id: cat_genmed, name: "General Medicine", localizedNames: { kn: "ಸಾಮಾನ್ಯ ವೈದ್ಯಕೀಯ", hi: "सामान्य चिकित्सा" } } ]
                      gender: FEMALE
                      price: { amount: 600, currency: INR, confirmed: true }
                      dataConfirmed: true
                    sessions:
                      - sessionId: ses_res_garima_2026-09-26_2
                        date: "2026-09-26"
                        label: Afternoon
                        start: "15:00"
                        end: "17:00"
                        status: SCHEDULED
                        timingCertainty: EXPECTED
                        presence: null
                        capacityModel: SEQUENCE
                        capacity: { total: 12, booked: 3, remaining: 9 }
                        arriveBy: "16:45"
                        bookable: true
                        notBookableReason: null
                        slots:
                          - { slotId: slot_ses_res_garima_2026-09-26_2_04, kind: SEQUENCE, position: 4, expectedWindow: { from: "15:30", to: "15:50" }, available: true }
                          - { slotId: slot_ses_res_garima_2026-09-26_2_05, kind: SEQUENCE, position: 5, expectedWindow: { from: "15:40", to: "16:00" }, available: true }
                          - { slotId: slot_ses_res_garima_2026-09-26_2_06, kind: SEQUENCE, position: 6, expectedWindow: { from: "15:50", to: "16:10" }, available: true }
                    unavailable: []
                alternatives: []
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /agent/bookings:
    post:
      tags: [Agent]
      operationId: agentCreateBooking
      summary: Reserve one slot for one customer
      description: >
        Atomically reserves the slot. RULE: if the slot was taken or the session
        changed since it was offered, return 409 SLOT_UNAVAILABLE with the
        session's current bookable slots so the agent can re-offer without a
        second search. RULE: one call per customer; two customers on one phone are
        two calls with two Idempotency-Keys. RULE: the response contains
        everything the agent needs to close the call (resource, date, window,
        arriveBy, confirmationCode). Status is BOOKED, which means a slot is
        held; timing remains an expectation unless timingCertainty=CONFIRMED.
      x-mcp:
        expose: true
        tool: manage_booking
        action: BOOK
        inject_headers: [X-Call-Id, X-Caller-Number, Idempotency-Key]
      parameters:
        - $ref: '#/components/parameters/CallId'
        - $ref: '#/components/parameters/CallerNumber'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentBookRequest' }
            example:
              slotId: slot_ses_res_garima_2026-09-26_2_04
              customer: { name: "Lakshmi R", phone: "9876543210", relationToCaller: SELF }
              reasonVerbatim: "ಮೊಣಕಾಲು ನೋವು ಎರಡು ವಾರದಿಂದ"
              language: kn
      responses:
        '201':
          description: Slot reserved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentBooking' }
              example:
                outcome: BOOKED
                bookingId: bkg_0192
                confirmationCode: "4471"
                status: BOOKED
                customer: { name: "Lakshmi R", phone: "9876543210" }
                resource: { resourceId: res_garima, name: "Dr. Garima", localizedNames: { kn: "ಡಾ. ಗರಿಮಾ" } }
                category: { id: cat_genmed, name: "General Medicine" }
                date: "2026-09-26"
                session: { sessionId: ses_res_garima_2026-09-26_2, label: Afternoon, start: "15:00", end: "17:00" }
                slot: { slotId: slot_ses_res_garima_2026-09-26_2_04, kind: SEQUENCE, position: 4, expectedWindow: { from: "15:30", to: "15:50" } }
                arriveBy: "16:45"
                timingCertainty: EXPECTED
                price: { amount: 600, currency: INR, confirmed: true }
                resourceToday: null
        '200':
          description: Idempotent replay.
          headers:
            Idempotent-Replay: { schema: { type: boolean } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentBooking' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/SlotConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
    get:
      tags: [Agent]
      operationId: agentListBookings
      summary: Bookings belonging to the caller's phone number
      description: >
        Scoped by X-Caller-Number (matches the booking phone or the number the
        booking was made from). RULE (tenant disclosure policy, default
        NAME_REQUIRED): if more than one customer is registered under the number
        and customerName is absent, return outcome=NAME_REQUIRED with
        customersOnNumber and no names. RULE: each item embeds the resource's
        current session state so "is my resource coming?" is one call.
      x-mcp:
        expose: true
        tool: manage_booking
        action: LIST
        inject_headers: [X-Call-Id, X-Caller-Number]
      parameters:
        - $ref: '#/components/parameters/CallId'
        - $ref: '#/components/parameters/CallerNumber'
        - name: customerName
          in: query
          schema: { type: string, maxLength: 100 }
        - name: phone
          in: query
          description: Only when the caller number is absent (landline) or the caller booked under another number. Read-only; cancel/reschedule on a spoken number is refused by default.
          schema: { $ref: '#/components/schemas/Phone' }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Bookings (possibly none).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentBookingList' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
  /agent/bookings/{bookingId}/cancel:
    post:
      tags: [Agent]
      operationId: agentCancelBooking
      summary: Cancel the caller's booking
      description: >
        RULE: succeeds only when X-Caller-Number matches the booking phone or the
        booked-from number AND customerName matches. Any mismatch returns the
        same neutral 404 as a non-existent booking. Releases the slot.
      x-mcp:
        expose: true
        tool: manage_booking
        action: CANCEL
        inject_headers: [X-Call-Id, X-Caller-Number, Idempotency-Key]
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/CallId'
        - $ref: '#/components/parameters/CallerNumber'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerName]
              properties:
                customerName: { type: string, maxLength: 100 }
                reasonVerbatim: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Cancelled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentBooking' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
  /agent/bookings/{bookingId}/reschedule:
    post:
      tags: [Agent]
      operationId: agentRescheduleBooking
      summary: Move the caller's booking to another slot
      description: >
        Same verification as cancel. newSlotId comes from find_availability and
        may belong to a different resource. Atomic: old slot released and new
        slot reserved together, or nothing changes (409 SLOT_UNAVAILABLE).
      x-mcp:
        expose: true
        tool: manage_booking
        action: RESCHEDULE
        inject_headers: [X-Call-Id, X-Caller-Number, Idempotency-Key]
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/CallId'
        - $ref: '#/components/parameters/CallerNumber'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerName, newSlotId]
              properties:
                customerName: { type: string, maxLength: 100 }
                newSlotId: { type: string }
      responses:
        '200':
          description: Rescheduled. Response includes previous slot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentBooking' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/SlotConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }

  # ===========================================================================
  # DIRECTORY — who exists, in every language the callers use
  # ===========================================================================
  /agent/knowledge-search:
    post:
      tags: [Agent]
      operationId: agentKnowledgeSearch
      summary: Answer a caller's general question from approved answers only
      description: >
        Deterministic retrieval over the knowledge base's question variants (any
        language or script, normalised and transliterated like the resolver).
        RULE: run the red-flag check first; if it fires, return
        routing.action=TRANSFER_EMERGENCY and nothing else.
        RULE: answer only above the tenant's answer threshold and clearly ahead
        of the runner-up; between the clarify and answer thresholds return
        CLARIFICATION_NEEDED with up to three options; below, NO_ANSWER. Never
        compose or translate an answer: `answer.language` says which approved
        text was returned.
        RULE: never NO_ANSWER for a failure; set outcome=COULD_NOT_CHECK.
        WHY: in healthcare an improvised answer is a risk; the provider signs off
        every sentence the agent may speak.
      x-mcp:
        expose: true
        tool: search_knowledge
        inject_headers: [X-Call-Id, X-Caller-Number]
      parameters:
        - $ref: '#/components/parameters/CallId'
        - $ref: '#/components/parameters/CallerNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgeSearchRequest' }
            example: { question: "ಪಾರ್ಕಿಂಗ್ ಇದೆಯಾ", language: kn }
      responses:
        '200':
          description: An approved answer, a clarification, or no answer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeSearchResponse' }
              example:
                outcome: ANSWERED
                asOf: "2026-09-26T10:02:11+05:30"
                routing: { action: ANSWER }
                answer:
                  entryId: kb_parking
                  topic: parking
                  text: "ಹೌದು. ನೆಲಮಾಳಿಗೆಯಲ್ಲಿ ಉಚಿತ ಪಾರ್ಕಿಂಗ್ ಇದೆ; ಮುಖ್ಯ ದ್ವಾರದ ಪಕ್ಕದ ರಸ್ತೆಯಿಂದ ಒಳಗೆ ಬನ್ನಿ."
                  language: kn
                  source: CURATED
                  confidence: 1.0
                  matchedQuestion: "ಪಾರ್ಕಿಂಗ್ ಇದೆಯಾ"
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Client credentials per consumer; tenant bound to the credential.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example-product.com/oauth2/token
          scopes:
            agent: Voice agent facade (all /agent operations)
            directory.write: Manage resources, categories, lexicon
            schedule.write: Manage templates and exceptions
            board.write: Update the live board
            bookings.staff: Staff booking operations, impact lists, notifications
            knowledge.write: Manage the knowledge base (approved answers)
            calls.write: Store call summaries
            calls.read: Read call summaries
  parameters:
    CallId:
      name: X-Call-Id
      in: header
      required: true
      description: Voice platform call id. Injected by the MCP layer; never a model argument.
      schema: { type: string, maxLength: 64 }
    CallerNumber:
      name: X-Caller-Number
      in: header
      required: false
      description: >
        Network caller number in E.164 from SIP, injected by the MCP layer.
        Absent for withheld or landline callers. Never a model argument.
      schema: { type: string, pattern: '^\+[1-9][0-9]{6,14}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Required on the agent path. Replay returns 200 + Idempotent-Replay. Retained 24 h.
      schema: { type: string, maxLength: 64 }
...
    BookingId:
      name: bookingId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    ValidationFailed:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Not found, or on verified operations the caller number or customer name did not match. Indistinguishable on purpose.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    SlotConflict:
      description: >
        SLOT_UNAVAILABLE: the slot was taken or the session changed since it
        was offered. details.currentSlots carries the session's current
        bookable slots so the agent can re-offer without another search.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: SLOT_UNAVAILABLE
              message: "Slot already booked"
              currentSlots:
                - { slotId: slot_ses_res_garima_2026-09-26_2_05, kind: SEQUENCE, position: 5, expectedWindow: { from: "15:40", to: "16:00" }, available: true }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Internal:
      description: Unexpected error. Nothing may be inferred about the data.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServiceUnavailable:
      description: Temporarily unavailable.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UpstreamTimeout:
      description: Write did not complete within 5 s. Retry once with the same Idempotency-Key; never with a new one.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Phone:
      type: string
      description: Contact number as dictated, digits only, national format for the tenant (10 digits for the first tenant). Distinct from X-Caller-Number.
      pattern: '^[0-9]{6,15}$'
      example: "9876543210"
...
    AvailabilitySearchRequest:
      type: object
      required: [utterance, language]
      description: >
        Everything optional except the caller's words. The LLM fills what it
        heard; the server does the understanding. Fields may be in any language
        or script. Nothing here is an id — the LLM never invents ids.
      properties:
        utterance:
          type: string
          maxLength: 500
          description: The caller's request verbatim (as STT delivered it).
        language: { $ref: '#/components/schemas/Language' }
        resourceName: { type: string, maxLength: 100, description: If the caller named a resource. As heard. }
        category: { type: string, maxLength: 100, description: If the caller named a speciality or category. As heard. }
        needText: { type: string, maxLength: 300, description: If the caller described a problem. As heard. Server maps via approved NEED_ROUTE only. }
        when:
          type: object
          description: Either an expression ("tomorrow evening", "ನಾಳೆ", "kal shaam") or explicit dates. Server resolves in facility time.
          properties:
            expression: { type: string, maxLength: 100 }
            dateFrom: { type: string, format: date }
            dateTo: { type: string, format: date }
            dayPart: { $ref: '#/components/schemas/DayPart' }
        preferences:
          type: object
          properties:
            gender: { $ref: '#/components/schemas/Gender' }
            language: { $ref: '#/components/schemas/Language' }
        maxResources: { type: integer, minimum: 1, maximum: 5, default: 3 }
        maxSlotsPerSession: { type: integer, minimum: 1, maximum: 5, default: 3 }
...
    AvailabilitySearchResponse:
      type: object
      required: [outcome, asOf, routing, understood, results, alternatives]
      properties:
        outcome:
          type: string
          enum: [FOUND, NONE_AVAILABLE, CLARIFICATION_NEEDED, TRANSFER, COULD_NOT_CHECK]
          description: The one field the agent branches on.
        asOf: { type: string, format: date-time }
        routing: { $ref: '#/components/schemas/Routing' }
        understood: { $ref: '#/components/schemas/Understood' }
        clarification: { $ref: '#/components/schemas/Clarification' }
        results:
          type: array
          items: { $ref: '#/components/schemas/ResourceResult' }
        alternatives:
          type: array
          description: Same-category resources with bookable sessions when the requested resource has none.
          items: { $ref: '#/components/schemas/ResourceResult' }
        partial:
          type: boolean
          description: True when part of the data could not be fetched; see notes.
        notes:
          type: array
          items: { type: string }
    AgentBookRequest:
      type: object
      required: [slotId, customer, language]
      properties:
        slotId: { type: string }
        customer:
          type: object
          required: [name, phone]
          properties:
            name: { type: string, maxLength: 100, description: As spoken and spelled back. }
            phone:
              allOf: [{ $ref: '#/components/schemas/Phone' }]
              description: Dictated and read back. Never defaulted from the caller number.
            relationToCaller:
              type: string
              enum: [SELF, CHILD, PARENT, SPOUSE, OTHER]
              nullable: true
        reasonVerbatim: { type: string, maxLength: 500, description: "Caller's words, any language. Never an interpretation." }
        language: { $ref: '#/components/schemas/Language' }
        requestTimingConfirmation:
          type: boolean
          default: false
          description: Caller wants the desk to confirm timing when the session is NOT_CONFIRMED. Honoured only if the tenant works that list.
    AgentBooking:
      type: object
      required: [outcome, bookingId, status, customer, resource, date, session, slot, timingCertainty]
      properties:
        outcome:
          type: string
          enum: [BOOKED, ALREADY_BOOKED, CANCELLED, RESCHEDULED, FOUND]
        bookingId: { type: string }
        confirmationCode:
          type: string
          description: Short code the agent reads back; lets a caller identify a booking on a later call without exposing ids.
        status: { $ref: '#/components/schemas/BookingStatus' }
        customer:
          type: object
          properties:
            name: { type: string }
            phone: { $ref: '#/components/schemas/Phone' }
        resource:
          type: object
          properties:
            resourceId: { type: string }
            name: { type: string }
            localizedNames: { $ref: '#/components/schemas/LocalizedText' }
        category: { $ref: '#/components/schemas/Category' }
        date: { type: string, format: date }
        session:
          type: object
          properties:
            sessionId: { type: string }
            label: { type: string }
            start: { $ref: '#/components/schemas/ClockTime' }
            end: { $ref: '#/components/schemas/ClockTime' }
        slot: { $ref: '#/components/schemas/Slot' }
        previousSlot:
          allOf: [{ $ref: '#/components/schemas/Slot' }]
          nullable: true
        arriveBy:
          allOf: [{ $ref: '#/components/schemas/ClockTime' }]
          nullable: true
        timingCertainty: { $ref: '#/components/schemas/TimingCertainty' }
        price:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          nullable: true
        followUp:
          type: string
          enum: [NONE, DESK_WILL_CONFIRM_TIMING]
          description: The agent may promise a desk call only when this is DESK_WILL_CONFIRM_TIMING.
        resourceToday:
          allOf: [{ $ref: '#/components/schemas/SessionInstance' }]
          nullable: true
          description: Present on LIST for same-day bookings — answers "is my resource coming?".
...